requirements.txt
asgiref==3.6.0
Babel==2.12.1
backports.zoneinfo==0.2.1
branca==0.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap4==2022.1
decorator==5.1.1
Django==4.2.1
django-cleanup==7.0.0
django-countries==7.5.1
django-crispy-forms==2.0
django-filter==23.2
django-phonenumber-field==7.1.0
folium==0.14.0
future==0.18.3
geocoder==1.38.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
phonenumberslite==8.13.11
Pillow==9.5.0
psycopg==3.1.9
pytz==2023.3
ratelim==0.1.6
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
